pyproject.toml
[project]
name = "plans-vision-api"
version = "1.0.0"
description = "Vision API for understanding construction plan conventions"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "openai>=1.57.0",
    "sqlalchemy>=2.0.36",
    "aiosqlite>=0.20.0",
    "greenlet>=3.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "structlog>=24.4.0",
    "httpx>=0.28.0",
    "tenacity>=9.0.0",
    "aiofiles>=24.1.0",
    "pillow>=11.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
